separate tile picker and brush editor
